feat(examples): add lofs of documentation and explanation for a simple example on schnorr dlog
